OCR/requirements.txt
pypdf==4.0.1
pdf2image==1.17.0
pytesseract>=0.3.13
pillow>=11.0.0
pandas>=2.2.0
numpy>=1.26.0
streamlit>=1.29.0
opencv-python>=4.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Production API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic[email]>=2.5.0
